Undo package restructure.
