+++ fabric-convention-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/convention/datagen/generators/EntityTypeTagGenerator.java
@@ -19,7 +19,7 @@ package net.fabricmc.fabric.impl.tag.convention.datagen.generators;
 import net.minecraft.entity.EntityType;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-import net.fabricmc.fabric.api.datagen.v1.provider.data.FabricTagProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
 import net.fabricmc.fabric.api.tag.convention.v1.ConventionalEntityTypeTags;
 
 public class EntityTypeTagGenerator extends FabricTagProvider.EntityTypeTagProvider {
