+++ fabric-data-generation-api-v1/src/main/java/net/fabricmc/fabric/api/datagen/v1/provider/FabricAdvancementProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.datagen.v1.provider.assets;
+package net.fabricmc.fabric.api.datagen.v1.provider;
 
 import java.io.IOException;
 import java.nio.file.Path;
