+++ fabric-data-generation-api-v1/src/main/java/net/fabricmc/fabric/mixin/datagen/ModelProviderMixin.java
@@ -42,7 +42,7 @@ import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-import net.fabricmc.fabric.api.datagen.v1.provider.assets.FabricModelProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricModelProvider;
 
 @Mixin(ModelProvider.class)
 public class ModelProviderMixin {
